# gin-tuto/docs/swagger.yaml
definitions:
  model.StGrafanaAlert:
    properties:
      dashboardId:
        type: integer
      evalMatches:
        items:
          $ref: '#/definitions/model.stEvalMatch'
        type: array
      imageUrl:
        type: string
      message:
        type: string
      orgId:
        type: integer
      panelId:
        type: integer
      ruleId:
        type: integer
      ruleName:
        type: string
      ruleUrl:
        type: string
      state:
        type: string
      tags:
        additionalProperties:
          type: string
        type: object
      title:
        type: string
    type: object
  model.StUniversalProducerEMail:
    properties:
      cc:
        example: xxx@yyyy.com;yyy@xxxx.co.kr
        type: string
      content:
        type: string
      from:
        type: string
      producer:
        example: wiss or wingo or kiss or more
        type: string
      tags:
        additionalProperties:
          type: string
        example:
          ' region': KR
          game: '#MUA2'
        type: object
      title:
        type: string
      to:
        example: xxx@yyyy.com;yyy@xxxx.co.kr
        type: string
    required:
    - content
    - from
    - producer
    - title
    - to
    type: object
  model.StUniversalProducerTeams:
    properties:
      content:
        type: string
      from:
        type: string
      producer:
        example: wiss or wingo or kiss or more
        type: string
      tags:
        additionalProperties:
          type: string
        example:
          ' region': KR
          game: '#MUA2'
        type: object
      title:
        type: string
      touri:
        example: http://xxx.x.xx.xxx.x.
        type: string
    required:
    - content
    - from
    - producer
    - title
    - touri
    type: object
  model.stEvalMatch:
    properties:
      metric:
        type: string
      tags:
        $ref: '#/definitions/model.stEvalMatchTags'
      value:
        type: integer
    type: object
  model.stEvalMatchTags:
    type: object
info:
  contact: {}
paths:
  /health:
    get:
      consumes:
      - application/json
      description: notifyhandler server heath check
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: site health check is running will return "working!"
  /notify/email:
    post:
      consumes:
      - application/json
      description: universal notify api for email
      parameters:
      - description: json struct for send email
        in: body
        name: jsonbody
        required: true
        schema:
          $ref: '#/definitions/model.StUniversalProducerEMail'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: 'universal Email api  : have just post api'
  /notify/grafana/email:
    post:
      consumes:
      - application/json
      description: 그라파나로 부터 메일을 통해 메세지를 보내고자할 때 사용 합니다
      parameters:
      - description: json struct for send email
        in: body
        name: jsonbody
        required: true
        schema:
          $ref: '#/definitions/model.StGrafanaAlert'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: 'Grafana mail api  : have just post api'
  /notify/grafana/teams:
    post:
      consumes:
      - application/json
      description: 그라파나로 부터 팀즈 를 향해 메세지를 보내고자 할 때 사용 합니다
      parameters:
      - description: json struct for send teams
        in: body
        name: jsonbody
        required: true
        schema:
          $ref: '#/definitions/model.StGrafanaAlert'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: 'Grafana teams api  : have just post api'
  /notify/teams:
    post:
      consumes:
      - application/json
      description: universal notify api for teams
      parameters:
      - description: json struct for send teams
        in: body
        name: jsonbody
        required: true
        schema:
          $ref: '#/definitions/model.StUniversalProducerTeams'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: 'Grafana teams api  : have just post api'
swagger: "2.0"